Add files via upload
The Tic Tac Toe project is about creating a fun two-player game where players take turns placing X's and O's on a 3x3 grid. The goal is to get three in a row. It’s a simple project that helps you practice coding, handling user input, and making interactive games.